Fixed label conversion order in loss calculation to ensure compatible data types. (#109)

+++ mipcandy/common/optim/loss.py
@@ -34,12 +34,12 @@ class DiceBCELossWithLogits(nn.Module):
         self.smooth: float = smooth
 
     def forward(self, masks: torch.Tensor, labels: torch.Tensor) -> tuple[torch.Tensor, dict[str, float]]:
-        labels = labels.float()
         if self.num_classes != 1 and labels.shape[1] == 1:
             d = labels.ndim - 2
             if d not in (1, 2, 3):
                 raise ValueError(f"Expected labels to be 1D, 2D, or 3D, got {d} spatial dimensions")
-            labels = convert_ids_to_logits(labels, d, self.num_classes)
+            labels = convert_ids_to_logits(labels.int(), d, self.num_classes)
+        labels = labels.float()
         bce = nn.functional.binary_cross_entropy_with_logits(masks, labels)
         masks = masks.sigmoid()
         soft_dice = soft_dice_coefficient(masks, labels, smooth=self.smooth)
